requirements.txt
alembic==1.7.1
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
cffi==1.14.6
cfgv==3.3.1
click==8.0.1
cryptography==3.4.8
distlib==0.3.2
ecdsa==0.17.0
fastapi==0.68.1
filelock==3.0.12
greenlet==1.1.1
h11==0.12.0
identify==2.2.14
Mako==1.1.5
MarkupSafe==2.0.1
nodeenv==1.6.0
passlib==1.7.4
platformdirs==2.3.0
pre-commit==2.15.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
python-decouple==3.4
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.23
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
virtualenv==20.7.2
